+++ OpenClosePrinciple/Geometrix/sources/Geometrix.WithOcp/Geometrix.WithOcp/ShapeModel/Triangle.cs
@@ -4,9 +4,9 @@ namespace iQuest.Geometrix.WithOcp.ShapeModel
 {
     internal class Triangle : IShape
     {
-        public int LengthA { get; set; }
-        public int LengthB { get; set; }
-        public int LengthC { get; set; }
+        public double LengthA { get; set; }
+        public double LengthB { get; set; }
+        public double LengthC { get; set; }
 
         public double CalculateArea()
         {
